requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.4
django-geoposition==0.3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.4
pytz==2018.4
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
